+++ lib/multi_filter_select_page.dart
@@ -3,26 +3,36 @@ import 'package:flutter/rendering.dart';
 import 'package:multiple_select/Item.dart';
 
 class MultiFilterSelectPage extends StatefulWidget {
-  final String placeholder;
+  final String? placeholder;
   final List<Item> allItems;
   final List initValue;
-  final bool autoFocusKeyboard;
+  final bool? autoFocusKeyboard;
   final bool searchCaseSensitive;
+  final Color selectedTextColor;
+  final Color selectedBackgroundColor;
+  final Color textColor;
+  final Color backgroundColor;
+  final Color borderColor;
 
   MultiFilterSelectPage(
       {this.placeholder,
-      @required this.searchCaseSensitive,
-      @required this.allItems,
-      @required this.initValue,
-      this.autoFocusKeyboard});
+      required this.searchCaseSensitive,
+      required this.allItems,
+      required this.initValue,
+      required this.autoFocusKeyboard,
+      required this.selectedTextColor,
+      required this.selectedBackgroundColor,
+      required this.textColor,
+      required this.backgroundColor,
+      required this.borderColor});
 
   @override
   State<StatefulWidget> createState() => MultiFilterSelectPageState();
 }
 
 class MultiFilterSelectPageState extends State<MultiFilterSelectPage> {
-  List<Item> filterItemList;
-  List selectedItemValueList;
+  late List<Item> filterItemList;
+  List? selectedItemValueList;
 
   @override
   void initState() {
@@ -38,14 +48,14 @@ class MultiFilterSelectPageState extends State<MultiFilterSelectPage> {
         appBar: AppBar(
           automaticallyImplyLeading: false,
           elevation: 0,
-          iconTheme: IconThemeData(color: Colors.black54),
+          iconTheme: IconThemeData(color: widget.textColor),
           backgroundColor: Colors.white,
           title: Row(
             children: <Widget>[
               GestureDetector(
                 child: Padding(
                   padding: EdgeInsets.only(right: 15),
-                  child: Icon(Icons.arrow_back_ios, size: 20),
+                  child: Icon(Icons.arrow_back_ios, size: 20, color: widget.textColor),
                 ),
                 onTap: () => Navigator.pop(context, this.selectedItemValueList),
               ),
@@ -56,27 +66,60 @@ class MultiFilterSelectPageState extends State<MultiFilterSelectPage> {
                     this.filterItemList = _filterItems(text);
                     this.setState(() {});
                   },
-                  autofocus: this.widget.autoFocusKeyboard,
+                  autofocus: this.widget.autoFocusKeyboard!,
                   style: TextStyle(fontSize: 14),
-                  cursorColor: Colors.grey,
+                  cursorColor: widget.textColor,
                   cursorWidth: 1.5,
                   decoration: InputDecoration(
-                      contentPadding:
-                          EdgeInsets.only(top: 0, bottom: 0, left: 10),
-                      fillColor: Colors.black12,
+                      contentPadding: EdgeInsets.only(top: 0, bottom: 0, left: 10),
+                      fillColor: widget.backgroundColor,
                       filled: true,
                       suffixIcon: Icon(
                         Icons.search,
                         size: 20,
-                        color: Colors.grey,
+                        color: widget.selectedTextColor,
                       ),
                       border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(25.0),
-                          borderSide: BorderSide.none),
+                          borderRadius: BorderRadius.circular(25.0), borderSide: BorderSide.none),
                       focusedBorder: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(25.0),
-                          borderSide: BorderSide.none),
+                          borderRadius: BorderRadius.circular(25.0), borderSide: BorderSide.none),
                       hintText: this.widget.placeholder ?? 'search……'),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(left: 8.0),
+                child: GestureDetector(
+                  onTap: () {
+                    for (var a in filterItemList) {
+                      setState(() {
+                        this.selectedItemValueList!.add(a.value);
+                      });
+                    }
+                  },
+                  child: Icon(
+                    Icons.add_circle_outline,
+                    size: 30,
+                  ),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(left: 8.0),
+                child: GestureDetector(
+                  onTap: () {
+                    for (var a in filterItemList) {
+                      bool _selected = this.selectedItemValueList!.contains(a.value);
+
+                      if (_selected) {
+                        setState(() {
+                          this.selectedItemValueList!.remove(a.value);
+                        });
+                      }
+                    }
+                  },
+                  child: Icon(
+                    Icons.remove_circle_outline,
+                    size: 30,
+                  ),
                 ),
               ),
             ],
@@ -92,14 +135,13 @@ class MultiFilterSelectPageState extends State<MultiFilterSelectPage> {
               runSpacing: 8.0,
               children: filterItemList.map(
                 (item) {
-                  bool _selected =
-                      this.selectedItemValueList.contains(item.value);
+                  bool _selected = this.selectedItemValueList!.contains(item.value);
                   return GestureDetector(
                     onTap: () {
                       if (_selected) {
-                        this.selectedItemValueList.remove(item.value);
+                        this.selectedItemValueList!.remove(item.value);
                       } else {
-                        this.selectedItemValueList.add(item.value);
+                        this.selectedItemValueList!.add(item.value);
                       }
                       this.setState(() {});
                     },
@@ -107,15 +149,14 @@ class MultiFilterSelectPageState extends State<MultiFilterSelectPage> {
                       padding: EdgeInsets.symmetric(vertical: 6, horizontal: 8),
                       child: Text(
                         item.content,
-                        style: TextStyle(fontSize: 15),
+                        style: TextStyle(
+                            fontSize: 15,
+                            color: _selected ? widget.selectedTextColor : widget.textColor),
                       ),
                       decoration: BoxDecoration(
-                        color: _selected ? Colors.blue[50] : Colors.grey[200],
+                        color: _selected ? widget.selectedBackgroundColor : widget.backgroundColor,
                         border: Border.all(
-                            width: 1,
-                            style: BorderStyle.solid,
-                            color:
-                                _selected ? Colors.blue[200] : Colors.black12),
+                            width: 1, style: BorderStyle.solid, color: widget.borderColor),
                         borderRadius: BorderRadius.circular(8.0),
                       ),
                     ),
@@ -134,11 +175,7 @@ class MultiFilterSelectPageState extends State<MultiFilterSelectPage> {
   }
 
   List<Item> _filterItems(String searchString) {
-    return this
-        .widget
-        .allItems
-        .where((item) => _itemInSearch(searchString, item))
-        .toList();
+    return this.widget.allItems.where((item) => _itemInSearch(searchString, item)).toList();
   }
 
   bool _itemInSearch(String searchString, Item item) {